# src/main/resources/swagger-ui/swagger.yaml
openapi: 3.0.3
info:
  title: MS Login
  description: >
    Microserviço responsável por autenticação, registro e gerenciamento de tokens JWT.
    Inclui endpoints para **login**, **registro**, **refresh** e **logout**.
  version: 1.0.1
...
paths:
...
  /refresh:
    post:
      summary: Renovar o token JWT
      description: >
        Gera um novo token JWT com base em um refresh token válido.
      operationId: refreshToken
      tags:
        - Autenticação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token renovado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
...
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Token de atualização JWT
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refreshToken...
...
    RefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: Token JWT gerado após login
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.accessToken...
        refreshToken:
          type: string
          description: Token de atualização
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refreshToken...
        expiresAt:
          type: string
          format: string
          description: Data e hora de expiração do token
          example: 2025-10-22T23:59:59Z
...
    ApiError:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Parâmetros inválidos.
        path:
          type: string
          example: /login
        timestamp:
          type: string
          format: date-time
          example: 2025-10-11T10:30:00Z
  responses:
    BadRequest:
      description: Requisição inválida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Falha de autenticação (token inválido ou expirado).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Acesso negado (usuário sem permissão).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InternalError:
      description: Erro interno do servidor.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'